replace totalRating with the sum of rating distribution in backend/put the files of use case 4 into folder StorePage
