--- frontend/src/components/Store/Store.css
@@ -1,66 +1,0 @@
-.left {
-    position: fixed;
-    left: 0;
-    top: 48px;
-    width: 20%;
-    height: 100%;
-    border-right: 3px solid #f44336;
-    padding-top: 5%;
-}
-
-.right {
-    position: absolute;
-    right: 0px;
-    width: 80%;
-    height: 100%;
-
-}
-
-
-.card-title {
-    background-color: #f44336;
-
-}
-
-.card-title h2 {
-    color: white;
-    font-size: 16px;
-    position: relative;
-    margin-top: 0px;
-    margin: 0px;
-}
-
-.card-title h3 {
-    font-size: 14px;
-    color: white;
-    margin: 0px;
-}
-
-
-
-
-.promote-icon {
-    color: 'white'
-}
-
-.icon2 {
-    position: relative;
-    float: right;
-    margin-right: 20px;
-    margin-bottom: 60px;
-}
-
-div.alef-card {
-    width: 100%;
-    height: 100%
-}
-
-div.alef-card-container {
-    width: 100%;
-    height: 60%;
-    background-color: #ffffff;
-}
-
-div.md-card-title {
-    height: 17%
-}
