--- frontend/src/components/StoreCard/CardFront.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react'
-export default class CardFront extends Component{
-    constructor(){
-        super()
-    }
-    render(){
-        return(
-            <img src='https://placeholdit.imgix.net/~text?txtsize=33&txt=318%C3%97180&w=318&h=180' style={{
-            width: '100%', height: '100%'}}/>
-        )
-    }
-
-}
