+++ frontend/src/components/StorePage/HorizontalChart.js
